requirements.txt
asgiref==3.6.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cmake==3.25.0
colorama==0.4.6
Django==3.2.16
django-cors-headers==3.13.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==6.0.0
numpy==1.21.6
opencv-python==4.7.0.68
Pillow==9.4.0
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
typing_extensions==4.4.0
whitenoise==6.3.0
zipp==3.11.0
